Move config from plugin to a separate module and make it to read the specific attributes from the CKAN config, add comments
